# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: JCode API
  version: 0.1.0
servers:
  - url: http://localhost:4000/api
paths:
...
  /escolas/{id}/turmas:
    post:
      summary: Criar turma (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '201': { description: Criado }
  /escolas/{id}/materias:
    post:
      summary: Criar matéria (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '201': { description: Criado }
  /escolas/{id}/usuarios:
    post:
      summary: Criar usuário (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '201': { description: Criado }
...
  /provas/{id}:
    put:
      summary: Atualizar prova (professor)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
...
  /uploads/imagem:
    post:
      summary: Upload de imagem (R2)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                prova_id:
                  type: string
                alt_text:
                  type: string
                legenda:
                  type: string
                prefer_glossary:
                  type: boolean
              required: [file, alt_text, prova_id]
      responses:
        '201': { description: Criado }
      description: |
        Aceita imagens (jpeg, png, webp, svg) e PDF. `prova_id` é obrigatório para gerar `ref_code` e vincular a imagem.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []